refactor(ccw): break down large files into focused modules under 250 lines
This refactoring significantly improves code organization and maintainability by
breaking down monolithic files into focused, single-responsibility modules.

## Refactoring Summary

### Major File Restructuring:
- **main.go**: 1141 → 146 lines (87% reduction)
  - Split into app/ package with 5 focused modules
  - Separated concerns: app.go, workflow.go, async_workflow.go, commands.go, converters.go

- **claude/claude.go**: 570 → removed (100% restructured)
  - Split into 5 focused modules: client.go, pr_description.go, implementation_summary.go, execution.go, context.go
  - Each module handles specific Claude integration functionality

- **types/models.go**: 539 → removed (100% restructured)
  - Split into 7 focused modules: github.go, workflow.go, validation.go, async.go, logging.go, performance.go, ui.go
  - Better type organization by domain

### Results:
- **30 files** now under 250 lines ✅
- **12 files** still over 250 lines (from 15 originally)
- **Improved maintainability** with focused single-responsibility modules
- **Better code organization** by functional domain
- **Easier navigation** and understanding of codebase
- **Zero functionality lost** - all features preserved

### Architecture Benefits:
- **Separation of Concerns**: Each file handles one specific responsibility
- **Modularity**: Related functionality grouped together
- **Testability**: Smaller modules are easier to test
- **Readability**: Developers can focus on relevant code
- **Maintainability**: Changes are isolated to specific domains

All builds pass and functionality is preserved while significantly improving code organization.
